requirements.txt
langchain-google-genai>=0.1.0
google-ai-generativelanguage>=0.1.0
google-generativeai
python-dotenv==1.0.1
Flask==2.3.3
pytesseract==0.3.10
Pillow==10.3.0
cryptography==43.0.0
requests
fuzzywuzzy
python-Levenshtein
pyzbar
opencv-python
numpy
flask-cors
langchain
langchain-community
faiss-cpu